book state added to BookForm
book state initialized once for BookForm component
checks to see if there is already a book in state
if there is the properties of the book object are
initialized to the properites of the state of the
incoming book. This will be useful later for editing.
If not properties are intialized to empty