handle input function complete, handleOnSubmit complete
handleOnSubmit checks to for the validitiy of each book property.
if they are all valid a new book constant is created from those properties
with two new properites, a unique id and a new date of creation.
If all fields are not valid the error state is updated and rendered